+++ Algorithm/Dijkastra.java
@@ -1,0 +1,71 @@
+import java.util.*;
+
+class Dijkstra {
+    private static final int V = 9; // Number of vertices in the graph
+
+    // A utility function to find the vertex with minimum distance value
+    int minDistance(int dist[], boolean sptSet[]) {
+        int min = Integer.MAX_VALUE, min_index = -1;
+
+        for (int v = 0; v < V; v++)
+            if (!sptSet[v] && dist[v] <= min) {
+                min = dist[v];
+                min_index = v;
+            }
+
+        return min_index;
+    }
+
+    // Function that implements Dijkstra's single source shortest path algorithm
+    void dijkstra(int graph[][], int src) {
+        int dist[] = new int[V]; // The output array. dist[i] will hold the shortest distance from src to i
+        boolean sptSet[] = new boolean[V]; // sptSet[i] will be true if vertex i is included in shortest path tree
+
+        // Initialize all distances as INFINITE and stpSet[] as false
+        for (int i = 0; i < V; i++) {
+            dist[i] = Integer.MAX_VALUE;
+            sptSet[i] = false;
+        }
+
+        // Distance of source vertex from itself is always 0
+        dist[src] = 0;
+
+        // Find shortest path for all vertices
+        for (int count = 0; count < V - 1; count++) {
+            // Pick the minimum distance vertex from the set of vertices not yet processed
+            int u = minDistance(dist, sptSet);
+
+            // Mark the picked vertex as processed
+            sptSet[u] = true;
+
+            // Update dist value of the adjacent vertices of the picked vertex
+            for (int v = 0; v < V; v++)
+                if (!sptSet[v] && graph[u][v] != 0 && dist[u] != Integer.MAX_VALUE && dist[u] + graph[u][v] < dist[v])
+                    dist[v] = dist[u] + graph[u][v];
+        }
+
+        // Print the constructed distance array
+        printSolution(dist);
+    }
+
+    // A utility function to print the constructed distance array
+    void printSolution(int dist[]) {
+        System.out.println("Vertex \t Distance from Source");
+        for (int i = 0; i < V; i++)
+            System.out.println(i + " \t\t " + dist[i]);
+    }
+
+    public static void main(String[] args) {
+        int graph[][] = new int[][] { { 0, 4, 0, 0, 0, 0, 0, 8, 0 },
+                                      { 4, 0, 8, 0, 0, 0, 0, 11, 0 },
+                                      { 0, 8, 0, 7, 0, 4, 0, 0, 2 },
+                                      { 0, 0, 7, 0, 9, 14, 0, 0, 0 },
+                                      { 0, 0, 0, 9, 0, 10, 0, 0, 0 },
+                                      { 0, 0, 4, 14, 10, 0, 2, 0, 0 },
+                                      { 0, 0, 0, 0, 0, 2, 0, 1, 6 },
+                                      { 8, 11, 0, 0, 0, 0, 1, 0, 7 },
+                                      { 0, 0, 2, 0, 0, 0, 6, 7, 0 } };
+        Dijkstra t = new Dijkstra();
+        t.dijkstra(graph, 0);
+    }
+}
