--- Heap_Sort.java
@@ -1,6 +1,0 @@
-package practice5;
-
-public class Heap_Sort
-{
-
-}
