+++ LinkedList/linkedlistcycle.java
@@ -1,0 +1,29 @@
+public class linkedlistcycle {
+    /**
+     * Definition for singly-linked list.
+     * class ListNode {
+     *     int val;
+     *     ListNode next;
+     *     ListNode(int x) {
+     *         val = x;
+     *         next = null;
+     *     }
+     * }
+     */
+
+
+        public boolean hasCycle(ListNode head) {
+            if(head==null) return false;
+            ListNode fast = head;
+            ListNode slow = head;
+
+            while(fast.next!=null && fast.next.next!=null)
+            {
+                fast=fast.next.next;
+                slow=slow.next;
+                if(fast==slow) return true;
+            }
+            return false;
+        }
+    }
+
