--- Mergetest1.java
@@ -1,4 +1,0 @@
-package practice5;
-
-public class Mergetest1 {
-}
